Add Routers and Pages

+++ backend/src/controllers/notesController.ts
@@ -2,10 +2,15 @@ import { RequestHandler } from 'express';
 import NoteModel from '../models/note';
 import createHttpError from 'http-errors';
 import mongoose from 'mongoose';
+import { assertIsDefined } from '../util/assertisDefined';
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
+
   try {
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(authenticatedUserId);
+
+    const notes = await NoteModel.find({ userId: authenticatedUserId }).exec();
     console.log('Server Response:', notes);
     res.status(200).json(notes);
   } catch (error) {
@@ -16,8 +21,10 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 export const getNote: RequestHandler = async (req, res, next) => {
   // To get the Id from the URL
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, 'Invalid note Id');
     }
@@ -27,6 +34,11 @@ export const getNote: RequestHandler = async (req, res, next) => {
     if (!note) {
       throw createHttpError(404, 'Note not found');
     }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, 'You cannot access this note');
+    }
+
     res.status(200).json(note);
   } catch (error) {
     next(error);
@@ -46,11 +58,16 @@ export const createNote: RequestHandler<
 > = async (req, res, next) => {
   const { title, text } = req.body;
 
+  const authenticatedUserId = req.session.userId;
+
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!title) {
       throw createHttpError(400, 'Note must have a title');
     }
     const newNote = await NoteModel.create({
+      userId: authenticatedUserId,
       title: title,
       text: text,
     });
@@ -82,7 +99,11 @@ export const updateNote: RequestHandler<
   const newTitle = req.body.title;
   const newText = req.body.text;
 
+  const authenticatedUserId = req.session.userId;
+
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, 'Invalid note Id');
     }
@@ -95,6 +116,10 @@ export const updateNote: RequestHandler<
 
     if (!note) {
       throw createHttpError(404, 'Note not found');
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, 'You cannot access this note');
     }
 
     note.title = newTitle;
@@ -112,7 +137,10 @@ export const updateNote: RequestHandler<
 export const deleteNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
 
+  const authenticatedUserId = req.session.userId;
+
   try {
+    assertIsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, 'Invalid note Id');
     }
@@ -121,6 +149,10 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
 
     if (!note) {
       throw createHttpError(404, 'Note not found');
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, 'You cannot access this note');
     }
 
     await NoteModel.deleteOne();
